step 4 -separated model and data logics to dif modules
